Completes sortBy
This is a rather inelegant way of doing this, especially after a bit of
looking at the actual underscore implementation. This is a situation where
the logic of underscore's "createCallback" helper function really becomes
apparent - I'd be interested in seeing what cases bring about some of the
additional checks in the underscore library, especially the || a === void 0
test in the 'sort' method and the initial transformation of the to be sorted
data into a mapped array of objects.