Fixes intersection and difference
Realized that if multiple copies of the same item exist in a given array, it
will be counted multiple times and thus break the algorithm - ie
['larry', 'moe', 'curly'], ['larry', 'larry', 'john'], ['john, 'george', 'paul']
would return ['larry'] because that entry was being counted twice in the second
passed array. Fixed by breaking the for loop once a match was found.